Prevent unnecessary tests for ClamAV and ClamSpam (#157)

+++ bin/fetch_clamspam.sh
@@ -75,6 +75,10 @@ fi
 
 ret=$(downloadDatas "$VARDIR/spool/clamspam/" "clamspam3" $randomize "clamav" "\|local_whitelist.ign2" "noexit")
 
+if [ ! -d "$VARDIR/spool/tmp/clamspam" ]; then
+	mkdir "$VARDIR/spool/tmp/clamspam"
+fi
+
 # Creating a test file, the content doesnt matter
 testfile=/tmp/scan-test.txt
 if [ ! -e ${testfile} ]; then
@@ -82,19 +86,35 @@ if [ ! -e ${testfile} ]; then
 fi
 
 # Getting to the ClamAV databases
-cd /var/mailcleaner/spool/clamav
+cd /var/mailcleaner/spool/clamspam
 # Foreach file
 for file in $(ls | grep -v "dbs.md5"); do
-#        echo "Checking ${file}"
-        # test if it is malformed
-        MALFORMEDFILE=`/opt/clamav/bin/clamscan -d ${file} ${testfile} 2>&1 > /dev/null |grep Malfor |grep -v ERROR | awk {'print $5'} |sed 's/:$//'`
 
-        # If the file is malformed, remove it
-        if [ ! -z ${MALFORMEDFILE} ] && [ -e ${MALFORMEDFILE} ] ; then
-                rm $MALFORMEDFILE
-                echo "["`date "+%Y/%m/%d %H:%M:%S"`"] Malformed Database $MALFORMEDFILE removed" >> /var/mailcleaner/log/mailcleaner/downloadDatas.log
-                MALFORMEDFILE=''
-        fi
+	# Check if we should re run test on this file
+	PERFORM_VERIFICATION=1
+	if [ -e "$VARDIR/spool/tmp/clamspam/$file" ]; then
+		CURRENT_MD5SUM=`md5sum $file |sed -e 's/ .*//'`
+		LAST_MD5SUM=`cat "$VARDIR/spool/tmp/clamspam/$file"`
+		if [ "$CURRENT_MD5SUM" = "$LAST_MD5SUM" ]; then
+			PERFORM_VERIFICATION=0
+		else
+			rm "$VARDIR/spool/tmp/clamspam/$file"
+		fi
+	fi
+
+        # test if it is malformed
+	if [ "$PERFORM_VERIFICATION" -eq "1" ]; then
+	        MALFORMEDFILE=`/opt/clamav/bin/clamscan -d ${file} ${testfile} 2>&1 > /dev/null |grep Malfor |grep -v ERROR | awk {'print $5'} |sed 's/:$//'`
+
+        	# If the file is malformed, remove it
+	        if [ ! -z ${MALFORMEDFILE} ] && [ -e ${MALFORMEDFILE} ] ; then
+        	        rm $MALFORMEDFILE
+                	echo "["`date "+%Y/%m/%d %H:%M:%S"`"] Malformed Database $MALFORMEDFILE removed" >> /var/mailcleaner/log/mailcleaner/downloadDatas.log
+	                MALFORMEDFILE=''
+		else
+			echo $CURRENT_MD5SUM > "$VARDIR/spool/tmp/clamspam/$file"
+	        fi
+	fi
 done
 cd -
